tools/promptbuilder/pyproject.toml
[project]
name = "promptbuilder"
version = "0.1.1" # Increment version
# ... other project metadata ...
requires-python = ">=3.8"
dependencies = [
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "pyperclip>=1.8", # Added rich
  "rich>=13.0",
  "prompt-toolkit>=3.0.51",
]

[project.scripts]
promptbuilder = "promptbuilder.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
